requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
accelerate>=0.27.0
peft>=0.10.0
datasets>=2.18.0
pyyaml>=6.0
wandb>=0.16.0
einops>=0.7.0
timm>=0.9.0
Pillow>=10.0.0
numpy>=1.24.0
